Fix bug in FsaeAuthorizationProvider
FsaeAuthorizationProvider had a bug where it tries to access jwtBody.role which is undefined. As a result any auth role checks would auto deny. Fixed so that .fsaeRole is accessed instead and renamed jwtBody to userProfile to better reflect that the field names are not the same as that of the JWT.